Implement exhaustive CI matrix and modular feature gating
### Overview
This PR addresses a critical gap in our CI/CD pipeline where seven optional domain modules (AI, Physics, Biology, Climate, etc.) were excluded from automated validation. Previously, the pipeline only verified the core library, leading to "hidden" regressions in specialized features that were often only caught during manual releases or at runtime.

The primary objective of these changes is to ensure 100% build integrity across the entire modular surface area, supporting both "core-only" and "all-features-enabled" configurations.

### Key Decisions & Rationale
- **Exhaustive Matrix Testing**: I implemented a GitHub Actions matrix that iterates through every domain module in isolation. This ensures that no module accidentally depends on another without an explicit declaration, maintaining strict modularity.
- **Feature-Aware Integration Tests**: Rather than having tests fail when a feature is disabled, I refactored the test suite using Rust's `#[cfg(feature = "...")]` attributes. This allows the test runner to dynamically include or exclude tests based on the active feature set.
- **Local-CI Parity**: To reduce the developer feedback loop, I've provided a local bash script that mirrors the CI environment. This allows contributors to verify their changes across all permutations before pushing code.
- **Dynamic Example Gating**: Updated `Cargo.toml` with `required-features` to prevent `cargo test` from attempting to compile examples that rely on inactive domains, which was a major source of build noise.

### Changes
#### CI/CD & Automation
- **GitHub Actions**: Appended a `feature-matrix` job to `.github/workflows/ci.yml` covering all seven modules plus a baseline.
- **Local Script**: Created `/app/scripts/test_features.sh` to facilitate local validation of modular combinations.

#### Engineering & Refactoring
- **Feature Gating**: Applied missing `#[cfg]` gates in `error.rs` and injected `#![cfg(all(...))]` attributes into over 30 integration test files.
- **Library Configuration**: Wrapped `math_explorer/src/lib.rs` testing blocks with combined feature checks to prevent compilation errors in minimal builds.
- **Cargo Cleanup**: Added `required-features` blocks to `Cargo.toml` for all domain-specific examples.

#### Bug Fixes (Identified via new testing)
- Fixed arithmetic incomplete states in the `pure_math` domain (specifically addressing `rug` library interactions).
- Corrected `biology` doctests by adding missing domain wrappers for `DiffusionCoeff` and `StepSize`.

### Verification Results
- [x] CI pipeline successfully reports status for all 7 modules.
- [x] Core-only builds (`cargo build --no-default-features`) now pass without error.
- [x] All Phase 8 TODOs regarding workspace-wide testing and WASM checks are resolved.
- [x] Local `test_features.sh` script executes successfully on a clean environment.

+++ math_explorer/src/pure_math/number_theory/ambs.rs
@@ -234,7 +234,7 @@ impl AmbsDsp {
                 }
             }
 
-            let check = (&x_n * &x_n).rem_euc(&p_k);
+            let check = Integer::from(&x_n * &x_n).rem_euc(&p_k);
             let a_mod = a.clone().rem_euc(&p_k);
             if check == a_mod {
                 final_roots.push(x_n);
@@ -364,8 +364,8 @@ impl AmbsDsp {
 
                 for (c, &is_valid) in c_valid.iter().enumerate().take(c_max) {
                     if is_valid {
-                        let z = &r_i + &(Integer::from(c) * &s_l);
-                        let z_sq = &z * &z;
+                        let z = Integer::from(&r_i + &Integer::from(Integer::from(c) * &s_l));
+                        let z_sq = Integer::from(&z * &z);
                         let _target_sigma = (Integer::from(2) * &n_l_int * &z_sq + Integer::from(1)) / &s_l;
                     }
                 }
